CF1/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-storage==1.32.0
ndjson